Extracted player preview into 2 components
